consumer/requirements.txt
kafka-python
psycopg2-binary
requests
pyspark
redis